requirements.txt
Flask==2.1.1
transformers==4.26.1
torch==1.13.1
psutil==5.9.0
faiss-cpu==1.7.2
numpy==1.21.5
google-auth==2.6.6
protobuf==3.19.4
